pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jira-outlook-env"
version = "0.1.0"
description = "Jira and Outlook search environment for OpenEnv"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.13.0",
    "bitsandbytes>=0.47.0",
    "datasets>=4.8.0",
    "fastapi>=0.110.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.0",
    "nbconvert>=7.17.0",
    "nbformat>=5.10.0",
    "openai>=1.51.0",
    "openenv-core[core]>=0.2.2",
    "peft>=0.19.0",
    "pydantic>=2.0.0",
    "torch==2.8.0",
    "torchaudio==2.8.0",
    "torchvision==0.23.0",
    "transformers>=4.56.0",
    "trl==1.2.0",
    "uvicorn>=0.30.0",
]
license = "MIT"

[project.optional-dependencies]
dev = [
    "build>=1.2.2",
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
py-modules = ["__init__", "action_rendering", "client", "inference", "models", "task_bank"]
packages = ["server"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
